workflow: add vote_for_disbural script
This combines the logic in manage_icp and the retrieve_proposal_canister_call.sh scripts
to improve the quality of information presented when voting specifically on ICP dibursals.

The existing voting script is generic and can apply to any type of method call, but this
flexibility also leaves it lacking in areas that are specific to ICP disbursals. The
aim with this new script is to address that shortfall. Shortfalls addressed are:
- Candid topic type is not decoded to text (so it's not easy to see what topic the propsoal is for)
- The subaccount balance isn't presented to the voter (which isn't essential information but is useful context)

Future improvements that could be made are presenting the current recipient accounts that are set up for that topic (can always check the code though)

Relevant prior commits are:
- 3f09e5865126fc8c142b0fc3ae7939ef44f529f4 (made the retrieve_proposal_canister_call.sh better at presenting ICP dibursals)
- 3a56f75fd19bdb7d6722fecb4e582fa65cc687fb (reverted the commit above because it impacted other proposals, in that case updating principals to include new members)